Added automatic versioning and deployment scripts to our Jenkinsfile.
This script can now automatically generate the version number and deploy the image to Docker. It also provides notification at the start and end of the build. Moreover, it has been updated to behave differently depending on whether current branch is 'master' or not. Applies to both local and remote development. This improves the CI/CD pipeline by removing manual labor, increasing the efficiency and reliability.