Upgrade csctracker-py-core version in requirements.txt
The version of csctracker-py-core was upgraded to 24.2.15 in the project's requirements.txt file. With this upgrade, certain dependencies in the ScheduleJobs.py file had to be adjusted for compatibility. Specifically, threading and datetime have been replaced with a more streamlined utility function for ID generation, resulting in a cleaner and more efficient code.